backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
openai==1.10.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
twilio==8.11.1
python-dotenv==1.0.0
alembic==1.13.1

# Testing dependencies
pytest==9.0.2
pytest-asyncio==1.3.0
httpx<0.28
pytest-cov==7.0.0
